### 2019-03-10 #1309 @whedon
[PRE REVIEW]: AutoClassWrapper: a Python wrapper for AutoClass C classification
**Submitting author:** @pierrepo (<a href="http://orcid.org/0000-0003-4177-3619">Pierre Poulain</a>)
**Repository:** <a href="https://github.com/pierrepo/autoclasswrapper" target ="_blank">https://github.com/pierrepo/autoclasswrapper</a>
**Version:** v1.4.1
