### 2019-03-10 #1309 @arfon
@pierrepo - this pull request should fix your paper: https://github.com/pierrepo/autoclasswrapper/pull/11
