### 2019-03-10 2a34793
Fix wrong or missing references
