+++ paper/paper.bib
@@ -196,3 +196,16 @@
   year = {2010},
   pages = {51-56}
 }
+
+@article{jimenez2017,
+  title = {Four Simple Recommendations to Encourage Best Practices in Research Software},
+  volume = {6},
+  issn = {2046-1402},
+  doi = {10.12688/f1000research.11407.1},
+  language = {en},
+  journal = {F1000Research},
+  author = {Jim\'enez, Rafael C. and Kuzak, Mateusz and Alhamdoosh, Monther and Barker, Michelle and Batut, B\'er\'enice and Borg, Mikael and {Capella-Gutierrez}, Salvador and Chue Hong, Neil and Cook, Martin and Corpas, Manuel and Flannery, Madison and Garcia, Leyla and Gelp\'i, Josep Ll. and Gladman, Simon and Goble, Carole and Gonz\'alez Ferreiro, Montserrat and {Gonzalez-Beltran}, Alejandra and Griffin, Philippa C. and Gr\"uning, Bj\"orn and Hagberg, Jonas and Holub, Petr and Hooft, Rob and Ison, Jon and Katz, Daniel S. and Lesko{\v s}ek, Brane and L\'opez G\'omez, Federico and Oliveira, Luis J. and Mellor, David and Mosbergen, Rowland and Mulder, Nicola and {Perez-Riverol}, Yasset and Pergl, Robert and Pichler, Horst and Pope, Bernard and Sanz, Ferran and Schneider, Maria V. and Stodden, Victoria and Suchecki, Rados\l{}aw and Svobodov\'a Va{\v r}ekov\'a, Radka and Talvik, Harry-Anton and Todorov, Ilian and Treloar, Andrew and Tyagi, Sonika and {van Gompel}, Maarten and Vaughan, Daniel and Via, Allegra and Wang, Xiaochuan and {Watson-Haigh}, Nathan S. and Crouch, Steve},
+  month = jun,
+  year = {2017},
+  pages = {876}
+}
