### 2019-03-10 #1309 @pierrepo
The proof is OK for me.
