### 2019-03-10 #1309 @pierrepo
Following @whedon recommendations, here is a list of potential reviewers (taken from this [list](https://bit.ly/joss-reviewers)):

- @rpetit3
- @jdeligt
- @a-slide
- @lowandrew
- @afrendeiro
