### 2019-03-11 #1309 @kyleniemeyer
Hi @pierrepo, we need to assign an editor to handle your submission before reviewers—hang tight!
