### 2019-04-08 61da560
Check column type before calculating stats
