### 2019-04-16 #1309 @trallard
Hi Pierre @pierrepo apologies for the long wait, let's get things moving now for you 😄

Dear potential reviewers (@rpetit3 @jdeligt @a-slide @lowandrew @afrendeiro)

You all seem to have valuable expertise that could be used in this review, I am wondering if you have the time and bandwidth to help os review this submission?

**Title**: AutoClassWrapper: a Python wrapper for AutoClass C classification
**Repo**: [https://github.com/pierrepo/autoclasswrapper](https://github.com/pierrepo/autoclasswrapper)

**Summary**: AutoClass is an unsupervised Bayesian classification system.
AutoClass C is an implementation of the AutoClass algorithm made by the NASA.
AutoClassWrapper is a Python wrapper to ease the use of Autoclass C.

Submitting author: @pierrepo Pierren Poulain
Version: 1.4.1
