### 2019-04-16 #1309 @rpetit3
Hi @trallard!

I would be glad to help reviewing @pierrepo's manuscript/software.

Full disclosure, this will be my first time formally reviewing a manuscript/software. To help get started I have read through the [Review Checklist](https://joss.readthedocs.io/en/latest/review_checklist.html), but I still might stumble over some things (I'll do my best not to!). So, any advice, feedback, and/or instructions would be very much welcomed.

Depending on reviewer turn out, I will definitely understand if you'd like to choose a more experienced reviewer.

Thank you!
Robert
