### 2019-04-16 #1309 @trallard
Hi @rpetit3  and @lowandrew  this is great! Do not worry at all, the review process for JOSS is very comprehensive and I will monitor the review regularly to make sure you need as much guidance/support as needed 😄

I am going to assign you as reviewers and see if any of the other folks are interested in reviewing before starting the process
