### 2019-04-17 #1390 @trallard
Hi  @rpetit3, @lowandrew!

Here is the official review issue. You will notice that each of you has a checklist with the items you need to go through.
I suggest using this issue to follow the review process and carry on any discussions.
If there are particular issues or something that needs more in-depth/longer conversations please open an issue on the software repository and reference it here so I can follow up.

If there are any issues you need help with or need clarification of the process at any point please ping me here 🙋🏻‍♀️
