### 2019-04-19 pierrepo/autoclasswrapper#12 @rpetit3
Add AutoClass C as a dependency
Hi Pierre!

I followed the [demo](https://github.com/pierrepo/autoclasswrapper/blob/master/demo/demo.ipynb) (successfully!) and it includes a few steps to install AutoClass C.

```
wget https://ti.arc.nasa.gov/m/project/autoclass/autoclass-c-3-3-6.tar.gz
tar zxvf autoclass-c-3-3-6.tar.gz
rm -f autoclass-c-3-3-6.tar.gz
export PATH=$PATH:$(pwd)/autoclass-c

# if you use a 64-bit operating system,
# you also need to install the standard 32-bit C libraries:
# sudo apt-get install -y libc6-i386
```

I would imagine some users might over look the Docs and demo. Considering this, It might be useful to also include these steps under *Installation and dependencies* in the [README](https://github.com/pierrepo/autoclasswrapper/blob/master/README.md).

Cheers,
Robert
