### 2019-04-19 pierrepo/autoclasswrapper#13 @rpetit3
Check AutoClass C is working
Hi Pierre,

I ran into the issue of also needing to install libc6-i386 for my AutoClass C installation. Without it, `autoclass` is recognized by `wrapper.search_autoclass_in_path()` but it never actually executes. Then `while not Path("autoclass-run-success").exists():` never passes.

I wonder if during `wrapper.search_autoclass_in_path()`, if maybe a simple execution of `autoclass` to extract the version information would could be added. Here's an example of `autoclass` output.

```
autoclass-c/autoclass

AUTOCLASS C (version 3.3.6unx)

 AutoClass Search:
      > autoclass -search <.db2[-bin] file path> <.hd2 file path>
             <.model file path> <.s-params file path>

 AutoClass Reports:
      > autoclass -reports <.results[-bin] file path> <.search file path>
             <.r-params file path>

 AutoClass Prediction:
      > autoclass -predict <test.. .db2 file path>
             <training.. .results[-bin] file path>
             <training.. .search file path> <training.. .r-params file path>
```

This check could even be implemented into `wrapper.Run()`, which could return the Run object if autoclass is valid, or return an exception if it isn't found or doesn't work.

Cheers,
Robert
