### 2019-04-19 #1390 @rpetit3
@pierrepo - Great work Pierre! AutoClassWrapper was easy to install and get started with. I think your work offers a valuable path to making AutoClass C more accessible to naive users (such as myself!). Another important aspect, is that AutoClassWrapper allows the user to easily integrate AutoClass C into reproducible Jupyter/IPython notebooks.

Here are a few suggestions

1. Missing installation instruction. On the main readme, AutoClass C isn't listed as a dependency. I think since it is required, it would be useful to either include the steps from the demo notebook for installing AutoClass C in the README.md file, or at minimum note that it is required in the README. I created an issue (https://github.com/pierrepo/autoclasswrapper/issues/12) for this.

**Important** For now this has prevented me from checking the **Installation instructions** box above.

2. A statement of need. Currently the Background, describes AutoClass C and its uses, then ends with AutoClass@IJM, a webserver, making AutoClass C more accessible. It doesn't say anything about a need for something like AutoClassWrapper, a wrapper for running AutoClass C locally. That is not to say there is not a need for AutoClassWrapper, there definitely is! I found only one other AutoClass C related package, written in R (see: https://github.com/wikiselev/autoclass-c). This to me suggests a need. In order to highlight this need, I think a sentence highlighting the lack of packages/wrappers for making AutoClass C more accessible locally is all that would be needed.

**Important** For now this as prevented me from checking the **A statement of need** box above.

3. Author Ordering. At the moment, Jean-Michel Camadro is listed as the first author.  This may have been intentional, but the citation at the bottom of the PDF reads *Camdro et al.*. I realize in most cases the actual citation would actually read *Camadro & Poulain*. But I thought it was worth pointing out since, based on the GitHub commits, you are the only code contributor. Please keep in mind, this is not in any way meant to negate any contributions Jean-Michel made. I just wanted to make sure it wasn't an unintended effect of the PDF creation.

4. I ran into a *user-error*, related to a non functioning AutoClass C installation, while going through the demo. I've gone into more details at https://github.com/pierrepo/autoclasswrapper/issues/13. This is just a feature suggestion, and should not affect publication.

5. I wonder if the last sentence of the background could be shortened?
```
Original:
To this aim, AutoClass@IJM (Achcar, Camadro, & Mestivier, 2009), a web server that utilizes
AutoClass C, has been published few years ago to make Bayesian classification more
accessible (see for instance results from (Simpson et al., 2011, Léger, Garcia, Ounissi,
Lelandais, & Camadro (2015), Elliott, Tanaka, Schwark, & Andrade (2018))).

Alternative:
To this aim, AutoClass@IJM (Achcar, Camadro, & Mestivier, 2009), a web server that utilizes
AutoClass C, has made Bayesian classification more accessible (Simpson et al., 2011, Léger,
Garcia, Ounissi, Lelandais, & Camadro (2015), Elliott, Tanaka, Schwark, & Andrade (2018))).
```

Again, great job with AutoClassWrapper, Pierre! Please feel free to discuss any of these suggestions!

Tania @trallard - Just a heads up, I will be starting paternity leave sometime in the near future. During this time, I might be a little slow to respond, but will try my best!
