### 2019-05-06 pierrepo/autoclasswrapper#14 @lowandrew
merge_dataframes() necessary even when only one dataframe is present?
Hi there,

I've been playing around with some toy datasets, and from what I can tell it's necessary to call merge_dataframes() even when only one dataframe is part of my dataset.

For example, calling the following ends up leading to an error `ERROR    'NoneType' object has no attribute 'to_csv'`. Additionally, when run as part of a script this error isn't actually raised, so downstream parts of the script try to run, making it more difficult to figure out the actual source of the problem. Adding in a `clust.merge_dataframes()` after adding input data fixes the problem, but it's not clear to me why I should have to merge dataframes if I only have one dataframe.

```python
import autoclasswrapper as wrapper
clust = wrapper.Input()
clust.add_input_data('test_data.tsv', 'real scalar', 0.01)
clust.create_db2_file()
```
