### 2019-05-06 pierrepo/autoclasswrapper#15 @lowandrew
API documentation - clarify required parameters.
Reading through the API documentation I'm often unclear on what parameters are required and what aren't - for example, for the `Input()` class the default value for `input_error` on `add_input_data()` is listed as `None`, so I assumed that I could leave the value blank. However, the following code ends up failing, and it was only through trial and error that I was able to pinpoint the source of the error. Having the API documentation modified so that it's more clear what parameters are required throughout and/or throwing errors when required parameters aren't entered would make it much easier for users to know what's going wrong.

```python
import autoclasswrapper as wrapper
import time

clust = wrapper.Input()
clust.add_input_data('test.tsv', 'real scalar')

clust.merge_dataframes()
clust.create_db2_file()
clust.create_hd2_file()
clust.create_model_file()
clust.create_sparams_file()
clust.create_rparams_file()

wrapper.search_autoclass_in_path()
run = wrapper.Run()
run.create_run_file()
run.run()
time.sleep(20)
results = wrapper.Output()
results.extract_results()
results.aggregate_input_data()
```
