### 2019-05-07 #1390 @lowandrew
@pierrepo @trallard I overall quite like the package and find it works fairly well, but I would like to see more added in the way of documentation/real world examples, as I found myself troubleshooting via trial and error more than I would have liked when playing around with the package.

1. As noted by @rpetit3, the installation instructions aren't quite complete - it's only in going through the demo notebook that the install instructions for Autoclass-C itself are found. Ideally I'd like to see the the binaries for Autoclass-C included in the python distribution so that everything is installed in one shot via PyPi, but adding instructions to the Install section of the README/readthedocs would also be acceptable.

2. While there is a demo jupyter notebook (which is great!), it only provided one toy example. Ideally I'd like to see some more fleshed out examples, as there appears to be lots of functionality in Autoclass that isn't covered by the toy example (something demonstrating what happens with missing data/discrete values, for example, would be excellent). It'd also be nice to see something pulled from a real-world dataset, even if it was downsampled heavily to make it more manageable. My other issue with the tutorial is that it showed what to do, but I was often missing out on why certain steps had to happen. As an example of that, I ran into problems when making my own dataset because I had it all in one dataframe and didn't call the `merge_dataframes()` method see (https://github.com/pierrepo/autoclasswrapper/issues/14). I was eventually able to get things to work, but more explanation of what each step does and why each step is necessary would be very welcome.

3. While the API does have documentation that appears to be fairly complete, I didn't find it to be particularly user friendly. As an example, see https://github.com/pierrepo/autoclasswrapper/issues/15. Throughout the API documentation, it isn't clear what parameters are required and which are optional, and when required parameters are missed errors don't always get thrown immediately, and so I was left having to reverse engineer at what point I'd missed something in a function call. Getting the API doc to be more clear throughout would be very helpful to users like myself that aren't familiar with Autoclass.

Overall though, I found that the wrapper made it fairly simple to get Autoclass up and running - with some more tutorial/documentation I think it makes for an excellent package. Sorry for the delay in getting the review done!
