### 2019-05-13 #1390 @pierrepo
Dear @rpetit3 & @lowandrew thank you for your reviews. I appreciate very much the time you spent to test and assess AutoClassWrapper.
You both noticed a lack in the documentation and I will try to improve this point.
Of course, I will also answer all the questions you raised.
If you agree I'll come back to you before mid-June (ping @trallard ).
