### 2019-06-17 86a96b4
Add AutoClass C installation procedure in README
Fix #12
