### 2019-06-17 pierrepo/autoclasswrapper#13 @pierrepo
Hi @rpetit3

that funny because I had the function `get_autoclass_version()` ready of this but did not use it much :-)
I will call this function before actually running the clustering.

Thanks for the advice!
