### 2019-06-17 dddac69
Check if AutoClass C is available before clustering
Fix #13
