+++ apt.txt
@@ -1,0 +1,4 @@
+# for 64-bit operating system,
+# install the standard 32-bit C libraries:
+libc6-i386
+
