+++ requirements.txt
@@ -1,0 +1,5 @@
+numpy
+pandas
+scipy
+matplotlib
+chardet
