### 2019-06-18 11c28bb
Keep Zenodo button on same line
