### 2019-06-18 9f0be60
Add notebook path into binder url
