### 2019-06-18 bee18aa
Remove WARNING from test message
