### 2019-06-18 afaaea4
Update lib32 dependencies
