### 2019-07-01 #1390 @pierrepo
Hello @kyleniemeyer
I'm on it. I've just fixed 2 issues opened by @rpetit3 ([#12](https://github.com/pierrepo/autoclasswrapper/issues/12) and [#13](https://github.com/pierrepo/autoclasswrapper/issues/13))
I will now handle those opened by @lowandrew (https://github.com/pierrepo/autoclasswrapper/issues).
