### 2019-07-01 pierrepo/autoclasswrapper#14 @pierrepo
Hi @lowandrew you're perfectly right. The `.merge_dataframe()` is missleading here.

I'll try to figure out something better.
