### 2019-07-01 a6048a1
Clarify prepare_input_data() method
Rename input.merge_dataframes() to prepare_input_data()

Fix #14
