+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 **Dev**
 - Print AutoClass C Version on run
 - Update jinja and urllib3 for security issues
-
+- Rename input.merge_dataframes() to input.prepare_input_data()
+- 
 **1.4.1**
 - Rename marker files to autoclass-run-success and autoclass-run-failure
 
