### 2019-07-01 f88f28d
Fix notebooks cleanup
