+++ doc/source/api_reference.rst
@@ -1,12 +1,0 @@
-API Reference
-=============
-
-This page provides documentation generated from the source files.
-
-.. toctree::
-   :maxdepth: 1
-
-   ./pages/input
-   ./pages/run
-   ./pages/output
-   ./pages/tools
