+++ doc/source/installation.rst
@@ -1,17 +1,0 @@
-Installation
-============
-
-``autoclasswrapper`` is available for Python 3.6+ only.
-
-It can be installed with ``pip``:
-
-.. code-block:: bash
-
-    $ python3 -m pip install --user autoclasswrapper
-
-or ``pipenv``:
-
-.. code-block:: bash
-
-    $ pipenv --three
-    $ pipenv install autoclasswrapper
