+++ doc/source/demo.rst
@@ -1,25 +1,0 @@
-Demo
-====
-
-You will find demo dataset and script in the ``demo`` directory of the project.
-
-To get the project, you can either clone it:
-
-.. code-block:: bash
-
-    $ git clone https://github.com/pierrepo/autoclasswrapper.git autoclasswrapper
-
-
-or download it:
-
-.. code-block:: bash
-
-    $ wget https://github.com/pierrepo/autoclasswrapper/archive/master.zip
-    $ unzip master.zip
-    $ mv autoclasswrapper-master/ autoclasswrapper
-
-
-.. toctree::
-   :maxdepth: 1
-
-   ./notebooks/demo
