### 2019-07-01 c0b2255
Create conda env files (minimal + dev)
