+++ environment.yml
@@ -1,0 +1,17 @@
+name: autoclasswrapper
+# channels priority
+# see https://bioconda.github.io/
+# conda-forge has highest prioriy
+channels:
+  - defaults
+  - bioconda
+  - conda-forge
+dependencies:
+  - python=3.7
+  - pip
+  - numpy
+  - matplotlib
+  - pandas
+  - scipy
+  - chardet
+  - jupyter
