### 2019-07-01 13aac52
Fix tests with input.merge_dataframes()
