### 2019-07-01 7ea7d27
Update / increase doc
Fix #15
