+++ autoclasswrapper/output.py
@@ -21,15 +21,15 @@ class Output():
 
     Parameters
     ----------
-    root_in_name : string (default: "autoclass")
+    root_in_name : string, optional (default: "autoclass")
         Root name to read input files generated by autoclass.
         Example: "autoclass" will lead to "autoclass.db2",
         "autoclass.model", "autoclass.s-params"...
-    root_out_name : string (default: "autoclass_out")
+    root_out_name : string, optional (default: "autoclass_out")
         Root name to write output files
         Ex.: "autoclass_out" will lead to "autoclass_out.cdt",
         "autoclass_out_stats.tsv"
-    tolerate_error : bool (default: False)
+    tolerate_error : bool, optional (default: False)
         If True, countinue generation of autoclass input files even if an
         error is encounter.
         If False, stop at first error.
@@ -179,7 +179,7 @@ class Output():
 
         Parameters
         ----------
-        with_proba : bool (default False)
+        with_proba : bool (default False), optional
             If True, also writes probability of case to belong to each class.
 
         """
