### 2019-07-01 82a28e6
Configure Read the Docs with yaml file
