+++ doc/autoclasswrapper.md
@@ -9,7 +9,7 @@ AutoClass is an unsupervised Bayesian classification system developed at the NAS
 
 [AutoClass C](https://ti.arc.nasa.gov/tech/rse/synthesis-projects-applications/autoclass/autoclass-c/) is the implementation of the AutoClass algorithm in C. It has been developed by Cheeseman and Stutz in 1996. AutoClass C has been successful in classifying data as diverse as infrared spectra of stars, protein structures, introns from human DNA sequences, Landsat satellites images, body pattern in the common cuttlefish, patterns between rich and poor countries, network traffic, or catchments in the Australian landscape. In proteomics and genomics, where thousands of proteins or genes are detected at once, AutoClass C has been proven to produce insightful results.
 
-However, AutoClass C user interface isn't very friendly and requires that data and parameters are input in a very precise way. To help user to prepare data, perform classification and analyze clusters, we developed AutoClassWrapper as a Python wrapper around AutoClass C.
+However, AutoClass C user interface isn't very friendly and requires that data and parameters are input in a very precise way. To help user to prepare input data, perform classification and analyze output clusters, we developed AutoClassWrapper as a Python wrapper around AutoClass C.
 
 
 # Installation
@@ -39,30 +39,30 @@ $ sudo apt-get install -y libc6-i386
 
 AutoClass C can handle different types of data:
 
-- *real scalar*: numerical values bounded by 0. Example: length, weight, age...
-- *real location*: numerical values, positive et negative. Example: position, microarray log ratio, elevation...
-- *discrete*: qualitative data. Example: color, phenotype, name...
+- *real scalar*: numerical values bounded by 0. Examples: length, weight, age...
+- *real location*: numerical values, positive et negative. Examples: position, microarray log ratio, elevation...
+- *discrete*: qualitative data. Examples: color, phenotype, name...
 
-Each data type must be entered as **separate input file**.
+Each data type must be entered in **separate input file** (one for each type).
 
-AutoClass C handles very well missing data. Missing data must be represented by nothing (no `NA`, `?`, `None`...).
+AutoClass C handles missing data. Missing data must be represented by nothing (no `NA`, `?`, `None`, `NULL`...).
 
-The usual workflow to prepare data is to instantiate an object from the class `Input()`:
+The usual workflow to prepare data is to instantiate an object from the `Input()` class:
 ```python
 import autoclasswrapper as wrapper
 clust = wrapper.Input()
 ```
 
-then add as many datasets as wanted, usually one per different data types (*real scalar*, *real location* and *discrete*):
+then add as many datasets as wanted, usually one per different data types:
 
 ```python 
 clust.add_input_data("example1.tsv", "real scalar")
 clust.add_input_data("example2.tsv", "real location")
 ```
 
-Prefered input data format is tab-separated values. User must provide the type of data. 
+Input data format is [tab-separated](https://en.wikipedia.org/wiki/Tab-separated_values) values. User must provide the type of data (either `real scalar`, `real location` or `discrete`). 
 
-The default error on real values is 0.01. Error is relative for *real scalar* (0.01 means 1%) but absolute for *real location*.
+The default error on real values is 0.01. Error is relative for *real scalar* values (0.01 means 1%) but absolute for *real location* values. There is no error for *discrete* values. For *real scalar* and *real location* values, custom error can be defined with the `input_error` parameter of the `.add_input_data()` method.
 
 The next step is then to prepare input data and generate input files required by AutoClass C:
 
@@ -77,31 +77,46 @@ clust.create_rparams_file()
 
 ## Classification / clustering 
 
-Once input files are created, one can run AutoClass C:
[…]
