### 2019-07-07 0ed332c
Remove unused dependency files for readthedocs
