+++ readthedocs_requirements.txt
@@ -1,9 +1,0 @@
-numpy
-pandas
-chardet
-scipy
-matplotlib
-sphinx
-sphinxcontrib-napoleon
-sphinx_rtd_theme
-nbsphinx
