+++ devtools/bumpversion.cfg
@@ -8,8 +8,6 @@ serialize =
 
 [bumpversion:file:autoclasswrapper/__init__.py]
 
-[bumpversion:file:doc/source/index.rst]
-
 [bumpversion:file:.zenodo.json]
 
 [bumpversion:file:CHANGELOG.md]
