### 2019-07-07 551f02b
Update instructions for release
