### 2019-07-07 e92bada
Remove pipenv
