### 2019-07-07 d2781c2
Update conda installation
