+++ Makefile
@@ -3,8 +3,8 @@ default: help
 init: ## Install miniconda
 	wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
 	bash miniconda.sh -b -p $HOME/miniconda
+	source $HOME/miniconda/etc/profile.d/conda.sh
 	hash -r
-	export PATH="$HOME/miniconda/bin:$PATH"
 	conda config --set always_yes yes --set changeps1 no
 	conda update -q conda
 	conda info -a
