### 2019-07-07 0d4e4da
Update Travis CI with conda
