+++ environment-dev.yml
@@ -7,7 +7,6 @@ channels:
   - bioconda
   - conda-forge
 dependencies:
-  - python=3.7
   - pip
   - numpy
   - matplotlib
