+++ Makefile
@@ -4,6 +4,7 @@ init: ## Install miniconda
 	wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
 	bash miniconda.sh -b -p ${HOME}/miniconda
 	. ${HOME}/miniconda/etc/profile.d/conda.sh
+	bash -c 'hash -r'
 	conda config --set always_yes yes --set changeps1 no
 	conda update -q conda
 	conda info -a
