### 2019-07-07 92acf2d
Add all steps of conda install
