### 2019-07-07 dd7b43e
Update dependency to install AutoClass C
