### 2019-07-07 01c1e07
Remove unnecessary travis_wait instruction
