### 2019-07-07 a771f41
Add quiet option to wget
