### 2019-07-07 d720dd8
Create dedicated conda env for readthedocs
