+++ .readthedocs.yml
@@ -22,4 +22,4 @@ python:
       path: .
 
 conda:
-  environment: environment-dev.yml
+  environment: environment-readthedocs.yml
