### 2019-07-10 2a89709
Add doc for input_separator_char parameter
