### 2019-07-10 996e4bf
Add sphinx-git in doc
