+++ environment-dev.yml
@@ -27,3 +27,4 @@ dependencies:
   - pip:
     - nbsphinx-link
     - recommonmark
+    - sphinx-git
