### 2019-07-11 7f849e9
Add more explanations on input data format
