### 2019-07-11 #1390 @pierrepo
Dear @rpetit3 and @lowandrew (cc @trallard),

once again, thank you both for your kind and constructive reviews.

Here are my comments and answers to your reviews.

---

@rpetit3 's review

>1. Missing installation instruction. On the main readme, AutoClass C isn't listed as a dependency

AutoClass C installation steps are now provided in:

- the main `README.md` file,
- the [documentation](https://autoclasswrapper.readthedocs.io/en/latest/),
- both demo notebooks.

> 2. A statement of need. Currently the Background, describes AutoClass C and its uses, then ends with AutoClass@IJM, a webserver, making AutoClass C more accessible. It doesn't say anything about a need for something like AutoClassWrapper, a wrapper for running AutoClass C locally.

Thanks for pointing this to me. This was so obvious to me that I totally forgot!

I added the following sentence at the end of the Background section :

*Albeit its proven efficiency and versatility, AutoClass C is not easy to configure and run locally for the end user. As far as we know of, there is only a [R wrapper ](https://github.com/wikiselev/autoclass-c) developed by M. Spivakov but with a limited interface.*

> 3. Author Ordering.

Jean-Michel Camadro was a key part in this project although he did not release any commit. Alphabetical order was chosen for author order. We both agreed this order is fine.

> 4. I ran into a user-error, related to a non functioning AutoClass C installation, while going through the demo

Thank you for pointing this error to me. It is now fixed (your issue has been closed).

> 5. I wonder if the last sentence of the background could be shortened?

Yes. Thank you! Nice suggestion.

---

@lowandrew 's review

> 1. As noted by @rpetit3, the installation instructions aren't quite complete

I'm very sorry for those very incomplete installation instructions regarding AutoClass C. As suggested also by @rpetit3, AutoClass C installation steps are now provided in README, documentation and demo notebooks.

> 2. Ideally I'd like to see some more fleshed out examples

I've created a second notebook with the analysis of the iris flowers dataset. I modified this *perfect* dataset to add discrete and missing values.

> 2. (continued) My other issue with the tutorial is that it showed what to do, but I was often missing out on why certain steps had to happen. As an example of that, I ran into problems when making my own dataset because I had it all in one dataframe and didn't call the merge_dataframes()

In the [online documentation](https://autoclasswrapper.readthedocs.io/en/latest/autoclasswrapper.html), I have added more explanations on what successive commands are doing and what are the most interesting options. I hope this helps.

The issue you raised about the `.merge_dataframe()` is closed.

> 3. Throughout the API documentation, it isn't clear what parameters are required and which are optional.

All non mandatory parameters are now clearly labeled as optional in the API documentation. Sorry for the confusion this brought to you.

---

I've  also created a convenient Binder link to ease quick overviews.
