### 2019-07-15 #1390 @rpetit3
Dear @pierrepo,

This is great! Thank you very much for considering my suggestions and making changes you saw fit. It is safe to say you have my approval for your manuscript.

Cheers!
Robert

@trallard - With the changes mentioned above, @pierrepo's manuscript has my approval for publication.
