### 2019-07-15 #1390 @lowandrew
@pierrepo

This looks good to me too! The new examples and documentation are excellent.

@trallard - this has my approval as well.
