### 2019-07-16 #1390 @trallard
🎉Thanks, for the review @rpetit3  and @lowandrew

@pierrepo  Can you please do the following before moving to publication?
- Create a new release of the package and report the version here
- Generate a new DOI in Zenodo and paste it here
- Do a last check on the paper for missing references, typos and the such?

Once completed we can move this forward
