### 2019-07-17 9a982d4
Create new release of the Journal of Open Source Software
