### 2019-07-17 #1390 @pierrepo
Hi @trallard Thanks!

> Create a new release of the package and report the version here

Release 1.5.1 has been created and is available [here](https://github.com/pierrepo/autoclasswrapper/tree/v1.5.1)

> Generate a new DOI in Zenodo and paste it here

The specific DOI for this release is 10.5281/zenodo.3339286
Please note there is also a *generic* DOI (for all releases) : 10.5281/zenodo.2527058

> Do a last check on the paper for missing references, typos and the such?

Sounds good to me :smile:
