### 2019-07-19 #1390 @trallard
@pierrepo  I have made a PR for a missing DOI, can you please merge for me to make some final checks before sending over to the editor in chief?
