### 2019-07-25 #1390 @trallard
@openjournals/joss-eics as the editor of this submission I would like to recommend this for acceptance at JOSS. I am now passing this onto you for the next steps
