### 2019-07-25 #1390 @kyleniemeyer
Hi @pierrepo, can you add your coauthor to the Zenodo archive? You should be able to just edit the metadata.
