### 2019-07-25 #1390 @kyleniemeyer
@pierrepo also, please merge the small PR I just submitted, which fixes some references in the paper: https://github.com/pierrepo/autoclasswrapper/pull/17
