### 2019-07-25 #1390 @pierrepo
Dear @trallard thank you for taking care of this manuscript until there :clap:

@kyleniemeyer PR has been accepted and coauthor has been added in the Zenodo archive:
https://zenodo.org/record/3339286
