### 2019-07-25 #1390 @kyleniemeyer
@pierrepo congrats on your paper's acceptance in JOSS! Thanks to @rpetit3 and @lowandrew for reviewing, and @trallard for editing!
