### 2019-07-25 #1390 @pierrepo
Dear @rpetit3 @lowandrew @trallard and @kyleniemeyer
Thank you very much for having helped me to improve this paper and eventually accepted it for publication in JOSS. This was a nice and smooth publishing experience. :clap: :tada:
